requirements-dev.txt
# Datalight
docopt
pyaml
jsonschema

# Zenodo
requests

# For documentation
numpy
matplotlib

